// src/typings/watermarkOptions.ts
enum WatermarkType {
  Text = "text",
  Image = "image",
}

enum FontFamilyType {
  Arial = "Arial",
  Monospace = "Monospace",
  Roboto = "Roboto",
  Georgia = "Georgia",
  Cursive = "Cursive",
}

enum TextColorType {
  White = "#fff",
  Black = "#000",
}

enum WatermarkPositionType {
  Center = "center",
  TopLeft = "top-left",
  TopRight = "top-right",
  BottomLeft = "bottom-left",
  BottomRight = "bottom-right",
}

export { WatermarkType, FontFamilyType, TextColorType, WatermarkPositionType };